move rest of apply tests
